refactor, styling, number formatting
