+++ src/components/atoms/Time.tsx
@@ -4,11 +4,13 @@ import { format, formatDistance } from 'date-fns'
 export default function Time({
   date,
   relative,
-  isUnix
+  isUnix,
+  className
 }: {
   date: string
   relative?: boolean
   isUnix?: boolean
+  className?: string
 }): ReactElement {
   const dateNew = isUnix ? new Date(Number(date) * 1000) : new Date(date)
   const dateIso = dateNew.toISOString()
@@ -19,6 +21,7 @@ export default function Time({
     <time
       title={relative ? format(dateNew, 'MMMM d, yyyy') : undefined}
       dateTime={dateIso}
+      className={className || undefined}
     >
       {relative
         ? formatDistance(dateNew, Date.now(), { addSuffix: true })
