+++ src/components/molecules/PoolTransactions.module.css
@@ -1,0 +1,3 @@
+.time {
+  color: var(--color-secondary);
+}
