+++ src/components/organisms/AssetActions/Pool/index.module.css
@@ -62,6 +62,10 @@
   margin-bottom: -2rem;
 }
 
+.transactions [class*='rdt_Pagination'] {
+  margin-bottom: -1rem;
+}
+
 @keyframes pulse {
   0% {
     background: transparent;
