+++ src/components/organisms/AssetActions/Pool/index.tsx
@@ -12,7 +12,7 @@ import EtherscanLink from '../../../atoms/EtherscanLink'
 import Token from './Token'
 import TokenList from './TokenList'
 import { graphql, useStaticQuery } from 'gatsby'
-import PoolTransactions from '../../PoolTransactions'
+import PoolTransactions from '../../../molecules/PoolTransactions'
 
 export interface Balance {
   ocean: number
@@ -283,12 +283,10 @@ export default function Pool({ ddo }: { ddo: DDO }): ReactElement {
             )}
           </div>
 
-          {hasAddedLiquidity && (
-            <div className={styles.transactions}>
-              <h3 className={styles.title}>Your Transactions</h3>
-              <PoolTransactions poolAddress={price?.address} minimal />
-            </div>
-          )}
+          <div className={styles.transactions}>
+            <h3 className={styles.title}>Your Pool Transactions</h3>
+            <PoolTransactions poolAddress={price?.address} minimal />
+          </div>
         </>
       )}
     </>
